backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simplyfi-por-platform"
version = "1.0.0"
description = "SimplyFI Platform for Proof of Reserves (PoR) Cryptocurrency Audit"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SimplyFI Team", email = "dev@simplyfi.com"}
]
keywords = [
    "cryptocurrency",
    "proof-of-reserves",
    "audit",
    "compliance",
    "blockchain"
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "pydantic-extra-types>=2.0.0",
    "prometheus-client>=0.18.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-prometheus>=0.41b0",
    "opentelemetry-exporter-jaeger-thrift>=1.20.0",
    "opentelemetry-exporter-otlp>=0.41b0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-instrumentation-celery>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "shap>=0.43.0",
    "matplotlib>=3.8.0",
    "joblib>=1.3.0",
    "PyYAML>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "pytest-timeout>=2.1.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "bandit[toml]>=1.7.5",
    "semgrep>=1.45.0",
    "pre-commit>=3.5.0",
    "ipython>=8.16.0",
    "jupyter>=1.0.0",
    "autopep8>=2.0.0"
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
    "sphinx-autodoc-typehints>=1.25.0",
    "myst-parser>=2.0.0"
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
py_version = 311
skip_glob = ["*/migrations/*"]
known_third_party = [
    "fastapi",
    "sqlalchemy",
    "pydantic",
    "redis",
    "celery",
    "opentelemetry",
    "prometheus_client",
    "numpy",
    "pandas",
    "sklearn",
    "scipy",
    "shap"
]
known_first_party = ["app"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101", "B601"]
tests = [
    "B201",
    "B301",
    "B302",
    "B303",
    "B304",
    "B305",
    "B306",
    "B307",
    "B308",
    "B309",
    "B310",
    "B311",
    "B312",
    "B313",
    "B314",
    "B315",
    "B316",
    "B317",
    "B318",
    "B319",
    "B320",
    "B321",
    "B322",
    "B323",
    "B324",
    "B325",
    "B401",
    "B402",
    "B403",
    "B404",
    "B405",
    "B406",
    "B407",
    "B408",
    "B409",
    "B410",
    "B411",
    "B412",
    "B413",
    "B501",
    "B502",
    "B503",
    "B504",
    "B505",
    "B506",
    "B507",
    "B601",
    "B602",
    "B603",
    "B604",
    "B605",
    "B606",
    "B607",
    "B608",
    "B609",
    "B610",
    "B611",
    "B612",
    "B701",
    "B702",
    "B703",
    "B704",
    "B705"
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--disable-warnings",
    "-ra"
]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests",
    "slow: slow tests",
    "asyncio: async tests"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod"
]
precision = 2
skip_covered = false

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict_optional = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
show_error_codes = true

plugins = ["sqlalchemy.ext.mypy.plugin"]

ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "alembic.*",
    "redis.*",
    "celery.*",
    "prometheus_client.*",
    "opentelemetry.*"
]
ignore_missing_imports = true

[tool.sqlalchemy]
sqlalchemy.future = true
